feat: load tutorial after authentication (#5090)
